refactor : Exist, NonExist Exception 리팩터링
