+++ src/main/java/nextstep/subway/common/exception/ControllerExceptionAdvice.java
@@ -1,8 +1,5 @@
-package nextstep.subway.common;
+package nextstep.subway.common.exception;
 
-import nextstep.subway.line.exception.ExistLineException;
-import nextstep.subway.line.exception.NonExistLineException;
-import nextstep.subway.station.exception.ExistStationException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,9 +11,15 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class ControllerExceptionAdvice {
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler({ExistLineException.class, NonExistLineException.class, ExistStationException.class})
-    public ResponseEntity<Void> exceptionHandler(RuntimeException e) {
+    @ExceptionHandler(ExistResourceException.class)
+    public ResponseEntity<Void> existResourceExceptionHandler(RuntimeException e) {
         return ResponseEntity.badRequest().build();
+    }
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(NonExistResourceException.class)
+    public ResponseEntity<Void> nonExistResourceExceptionHandler(RuntimeException e) {
+        return ResponseEntity.notFound().build();
     }
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
