+++ src/main/java/nextstep/subway/common/exception/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common;
+package nextstep.subway.common.exception;
 
 public class ExceptionMessage {
