+++ src/main/java/nextstep/subway/common/exception/ExistResourceException.java
@@ -1,0 +1,8 @@
+package nextstep.subway.common.exception;
+
+public class ExistResourceException extends RuntimeException {
+
+    public ExistResourceException(String message) {
+        super(message);
+    }
+}
