+++ src/main/java/nextstep/subway/common/exception/NonExistResourceException.java
@@ -1,0 +1,8 @@
+package nextstep.subway.common.exception;
+
+public class NonExistResourceException extends RuntimeException {
+
+    public NonExistResourceException(String message) {
+        super(message);
+    }
+}
