--- src/main/java/nextstep/subway/line/exception/ExistLineException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.line.exception;
-
-public class ExistLineException extends RuntimeException {
-
-    public ExistLineException(String message) {
-        super(message);
-    }
-}
