--- src/main/java/nextstep/subway/line/exception/NonExistLineException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.line.exception;
-
-public class NonExistLineException extends RuntimeException {
-
-    public NonExistLineException(String message) {
-        super(message);
-    }
-}
