+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,17 +1,17 @@
 package nextstep.subway.station.application;
 
+import nextstep.subway.common.exception.ExistResourceException;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
 import nextstep.subway.station.dto.StationRequest;
 import nextstep.subway.station.dto.StationResponse;
-import nextstep.subway.station.exception.ExistStationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static nextstep.subway.common.ExceptionMessage.EXCEPTION_MESSAGE_EXIST_STATION_NAME;
+import static nextstep.subway.common.exception.ExceptionMessage.EXCEPTION_MESSAGE_EXIST_STATION_NAME;
 
 @Service
 @Transactional
@@ -32,7 +32,7 @@ public class StationService {
 
     private void validateStationName(String stationName) {
         if (stationRepository.existsByName(stationName)) {
-            throw new ExistStationException(EXCEPTION_MESSAGE_EXIST_STATION_NAME);
+            throw new ExistResourceException(EXCEPTION_MESSAGE_EXIST_STATION_NAME);
         }
     }
