--- src/main/java/nextstep/subway/station/exception/ExistStationException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.station.exception;
-
-public class ExistStationException extends RuntimeException {
-
-    public ExistStationException(String message) {
-        super(message);
-    }
-}
